complete

+++ src/three/perspectivecamera.rs
@@ -6,7 +6,7 @@ extern "C" {
 
     pub type PerspectiveCamera;
     #[wasm_bindgen(constructor)]
-    pub fn new(fov: f32, aspect: f32, near: f32, far: f32) -> PerspectiveCamera;
+    pub fn new(fov: f64, aspect: f64, near: f64, far: f64) -> PerspectiveCamera;
     #[wasm_bindgen(method, getter = position)]
     pub fn position(this: &PerspectiveCamera) -> Vector3;
